* Reusability example :
